requirements.txt
requests
django-filter
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
